requirements.txt
anyio==3.3.4
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
black==21.9b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
euclid3==0.1
idna==3.3
ipydatawidgets==4.2.0
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
json5==0.9.6
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prettytable==0.7.2
prometheus-client==0.12.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pypng==0.0.19
pyrsistent==0.18.0
python-dateutil==2.8.2
pythreejs==2.3.0
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.23
requests==2.26.0
requests-unixsocket==0.2.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
solidpython==1.1.1
svgpathtools==1.4.2
svgwrite==1.4.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
traittypes==0.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
viewscad==0.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
